python_gpt_po/requirements.txt
polib==1.2.0
openai==v1.3.6
python-dotenv==1.0.0.
